pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "flask-grade-tracker"
version = "1.0.0"
description = "A Flask-based grade tracking application with Canvas integration"
authors = [
    {name = "Grade Tracker Team"}
]
requires-python = ">=3.11"
dependencies = [
    # Core Flask Framework
    "Flask==2.3.3",
    "Werkzeug==2.3.7",
    
    # Database
    "Flask-SQLAlchemy==3.0.5",
    "PyMySQL==1.1.1",
    "Flask-Migrate==4.0.5",
    
    # Authentication & Security
    "Flask-Login==0.6.3",
    "Flask-WTF==1.1.1",
    "cryptography==44.0.1",
    
    # Caching & Performance
    "Flask-Caching==2.1.0",
    "redis>=4.6.0",
    "Flask-Limiter==3.5.0",
    
    # Background Tasks
    "celery>=5.3.0",
    "kombu>=5.3.0",
    
    # Canvas Integration
    "requests>=2.31.0",
    "canvasapi>=3.2.0",
    
    # Analytics & ML
    "pandas>=2.1.0",
    "numpy>=1.24.0",
    "scikit-learn>=1.3.0",
    
    # Email
    "Flask-Mail==0.9.1",
    
    # Environment & Configuration
    "python-dotenv>=1.0.0",
    
    # Production WSGI server
    "gunicorn==21.2.0",
    "gevent==23.9.1",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-cov",
    "flake8",
    "black",
    "isort",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]

[tool.black]
line-length = 88
target-version = ['py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88